Refactor: Database restructure for Entries
This branch modifies the structure of the Filesystem Storage database. The root directory of Filesystem Storage is renamed from 'secrets' to 'filesystem_storage' for clarity. Additionally, a new tab-specific folder named 'vaults' is expected to store all vaults-related files in a dedicated directory ('filesystem_storage/vaults'). Functionally, the 'filesystem_storage/vaults' tab folder becomes the main directory for the whole application component dedicated to vaults.

Dedicated tab folders prevent any future conflicts of groups of items of various types which would otherwise have to be stored together in the root directory ('filesystem_storage').

Isar database is mostly not affected by this branch. The only practical consequence is that the file path in FilesystemPathString field in entities will now include the new tab directory 'vaults'.

There is no database backwards compatibility at this stage.

List of changes:
- updated secrets_service.dart to create tab Filesystem Storage file ('vaults.snggle') together with the first created vault
- updated filesystem_storage_manager.dart to delete tab Filesystem Storage file ('vaults.snggle') together with the last deleted vault
- renamed filesystem_storage_key.dart to filesystem_storage_root_dir.dart, and renamed the Filesystem Storage root directory name from 'secrets' to `filesystem_storage`
- updated app_service.dart, changing isDataBaseExist() method to search the 'filesystem_storage' directory instead of the root 'secrets'
- updated secrets_repository.dart and filesystem_storage_manager.dart, adding the isSecretExists() method to check for existence of a Filesystem Storage file
- added filesystem_storage_tab_dir.dart enum for storing names of subdirectories within 'filesystem_storage'
- updated vault_list_page.dart, redefining vaults location from 'secrets' to the 'filesystem_storage/vaults' directory
- updated mock databases to follow the new structure

+++ lib/infra/managers/filesystem_storage/filesystem_storage_manager.dart
@@ -60,7 +60,8 @@ class FilesystemStorageManager {
 
     Directory parentDirectory = await _getParentDirectory(filesystemPath);
     bool parentDirectoryEmptyBool = parentDirectory.listSync().isEmpty;
-    if (parentDirectoryEmptyBool) {
+    bool parentDirectoryTopLevelBool = parentDirectory.path.endsWith('/vaults');
+    if (parentDirectoryEmptyBool && parentDirectoryTopLevelBool == false) {
       await parentDirectory.delete();
     }
   }
